Version 1.1.51 - Now the reactant-centered model resets the "alternating delta" variables once it has found an oscillation. This should avoid unpleasant non-determinism - EdgeDialog contains some (commented) code so that if we find a good way/heuristic to translate the k among different scenarios to obtain (more or less) similar behaviors, we will just need to add the conversions - InatResultPanel now has more precise null-checks on Cytoscape networks and views (there is a special object for both to represent "null"): hopefully this will help us understand why some models contain multiple copies of the same "Network 0" (??). Also, we delete the network&vizmap associated to a panel when the panel closes only if we are still in the same session where the panel was created. Finally, the InatResultPanel will make the Cytoscape Results panel "docked" only if it was hidden: if the user likes to keep that panel floating we no longer interfer. - FileUtils was udpated so that the rendering of a JPEG image would ask also for the zoom level (the one determining graph line thickness and font size) and image quality (in [0, 1], with 1 highest quality). At the end of the operation, a message is shown reporting success or failure. - consequently to the modification in FileUtils, the Graph was updated so that the zoom variable (named SCALA) would be accessible through get/set functions. Also, the zoom is passed on to the Series in the graph, so that they can properly scale the horizontal lines defining the standard deviations. - Series and OverlaySeries classes: modification to properly scale the horizontal line for StdDev as said just above

+++ src/inat/cytoscape/InatResultPanel.java
@@ -141,10 +141,10 @@ public class InatResultPanel extends JPanel implements ChangeListener, GraphScal
 		Heptuple<Model, SimpleLevelResult, Double, String, String, HashMap<String, HashMap<String, Object>>, HashMap<String, HashMap<String, Object>>> simulationData = loadSimulationData(simulationDataFile, false);
 		InatResultPanel panel = new InatResultPanel(simulationData.first, simulationData.second, simulationData.third, simulationData.fourth, null);
 		panel.savedNetwork = Cytoscape.getNetwork(simulationData.fifth);
-		if (panel.savedNetwork != null) {
+		if (panel.savedNetwork != null && !panel.savedNetwork.equals(Cytoscape.getNullNetwork())) {
 			//System.err.println("La rete salvata " + panel.savedNetwork.getIdentifier() + " ha " + panel.savedNetwork.getNodeCount() + " nodi e " + panel.savedNetwork.getEdgeCount() + " edgi");
 			CyNetworkView savedNetworkView = Cytoscape.getNetworkView(simulationData.fifth);
-			if (savedNetworkView != null) { //Keep all saved networks hidden
+			if (savedNetworkView != null && !savedNetworkView.equals(Cytoscape.getNullNetworkView())) { //Keep all saved networks hidden
 				Cytoscape.destroyNetworkView(savedNetworkView);
 				CyNetworkView otherView = null;
 				Collection<CyNetworkView> c = Cytoscape.getNetworkViewMap().values();
@@ -192,7 +192,7 @@ public class InatResultPanel extends JPanel implements ChangeListener, GraphScal
 		this.result = result;
 		this.scale = scale;
 		this.title = title;
-		if (originalNetwork != null) {
+		if (originalNetwork != null && !originalNetwork.equals(Cytoscape.getNullNetwork())) {
 			try {
 				CyNetworkView view = Cytoscape.getNetworkView(originalNetwork.getIdentifier());
 				@SuppressWarnings("rawtypes")
@@ -868,7 +868,7 @@ public class InatResultPanel extends JPanel implements ChangeListener, GraphScal
 	
 	
 	public void closeResultsPanel(CytoPanel cytoPanel) {
-		if (savedNetwork != null) {
+		if (savedNetwork != null && countSessionChanges <= 1) { //We destroy network & vizmap only if we are still in "our" session. Otherwise, we would risk destroying a network from another session
 			try {
 				final VisualMappingManager vizMap = Cytoscape.getVisualMappingManager();
 				CalculatorCatalog visualStyleCatalog = vizMap.getCalculatorCatalog();
@@ -1073,7 +1073,9 @@ public class InatResultPanel extends JPanel implements ChangeListener, GraphScal
 		container.add(buttons, BorderLayout.NORTH);
 		
 		
-		cytoPanel.setState(CytoPanelState.DOCK);
+		if (cytoPanel.getState().equals(CytoPanelState.HIDE)) {
+			cytoPanel.setState(CytoPanelState.DOCK); //We show the Results panel if it was hidden.
+		}
 		fCytoPanel = (CytoPanelImp)cytoPanel;
 		cytoPanel.addCytoPanelListener(this);
